Get rid of new file, now that code is simpler
